% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshAddPoint}
\alias{antsrMeshAddPoint}
\title{antsrMeshAddPoint}
\usage{
antsrMeshAddPoint(mesh, point, identifier = NA)
}
\arguments{
\item{mesh}{an 'antsrMesh'}

\item{point}{spatial point to add to mesh}
}
\description{
add point to mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshAddPoint( x, c(0,0,0) )
}
