% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshGetCell}
\alias{antsrMeshGetCell}
\title{antsrMeshGetCell}
\usage{
antsrMeshGetCell(mesh, identifier)
}
\arguments{
\item{mesh}{an 'antsrMesh'}

\item{identifier}{identifier of point to get}
}
\description{
get a given cell in mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshAddPoint( x, c(0,0,0) )

}
