% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshGetNumberOfCells}
\alias{antsrMeshGetNumberOfCells}
\title{antsrMeshGetNumberOfCells}
\usage{
antsrMeshGetNumberOfCells(mesh)
}
\arguments{
\item{mesh}{an 'antsrMesh'}
}
\description{
get number of cells in an antsrMesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshGetNumberOfCells(x)
}
