% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshGetNumberOfPoints}
\alias{antsrMeshGetNumberOfPoints}
\title{antsrMeshGetNumberOfPoints}
\usage{
antsrMeshGetNumberOfPoints(mesh)
}
\arguments{
\item{mesh}{an 'antsrMesh'}
}
\description{
get number of points in an antsrMesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshGetNumberOfPoints(x)
}
