% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshGetPoint}
\alias{antsrMeshGetPoint}
\title{antsrMeshGetPoint}
\usage{
antsrMeshGetPoint(mesh, identifier)
}
\arguments{
\item{mesh}{an 'antsrMesh'}

\item{identifier}{identifier of point to get}
}
\description{
get a given point in mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshSetPoint( x, 0, c(0,0,0) )
pt = antsrMeshGetPoint(x, 0)
}
