% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshGetPoints}
\alias{antsrMeshGetPoints}
\title{antsrMeshGetPoints}
\usage{
antsrMeshGetPoints(mesh, identifiers = NULL)
}
\arguments{
\item{mesh}{an 'antsrMesh'}

\item{identifiers}{identifiers of points to get}
}
\description{
get all points in mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshAddPoint( x, c(0,0,0) )
pt = antsrMeshGetPoints(x, 0)
}
