% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrRegistration_class.R
\docType{class}
\name{antsrRegistration-class}
\alias{antsrRegistration-class}
\alias{show,antsrRegistration-method}
\title{antsrRegsitration}
\usage{
\S4method{show}{antsrRegistration}(object)
}
\arguments{
\item{object}{input object to convert}
}
\description{
C++ type used to represent an ITK image registration filter
}
\section{Slots}{

\describe{
\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{precision}}{math precision is float or double'}

\item{\code{components}}{number of components in image}

\item{\code{pixeltype}}{pixeltype of images}

\item{\code{smoothing}}{smoothing for each level}

\item{\code{transform}}{type of transform}

\item{\code{optimizer}}{optimizer type to use}

\item{\code{interpolator}}{interpolator type to use}

\item{\code{sampling}}{sampling type for image-to-image metric}

\item{\code{metric}}{image-to-image metric for registration}

\item{\code{iterations}}{max iterations at each level}

\item{\code{shrink}}{shrink factor for each level}

\item{\code{samplingStrategy}}{type of metric sampling}

\item{\code{convergence}}{metric convergence value}

\item{\code{smoothingPhysical}}{flag indicating smoothing values are in mm (default in in voxels)}

\item{\code{nLevels}}{number of levels in registration}
}}

