% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrRegistration_class.R
\name{antsrRegistrationRun}
\alias{antsrRegistrationRun}
\alias{initialize,antsrRegistration-method}
\title{antsrRegistrationCreate}
\usage{
antsrRegistrationRun(stages, fixed, moving)
}
\arguments{
\item{dimension}{number of dimensions}

\item{percision}{use 'float' or 'double' for values}

\item{reserve}{number of points to allocate on creation}
}
\value{
antsrRegistration
}
\description{
create object that registers two images
}
\examples{
x =  antsrRegistrationCreate( 3, "float", reserve=128 )
}
