% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{applyAntsrTransformToImage}
\alias{applyAntsrTransformToImage}
\title{applyAntsrTransformToMesh}
\usage{
applyAntsrTransformToImage(transform, image, reference,
  interpolation = "linear")
}
\arguments{
\item{transform}{antsrTransform}

\item{image}{antsrMesh to transform}

\item{interpolation}{type of interpolator to use}
}
\value{
antsImage
}
\description{
Apply transform/s to an antsrMesh
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsrTransform", precision="float", type="AffineTransform", dimension=2 )
setAntsrTransformParameters(tx, c(0.9,0,0,1.1,10,11) )
img2 = applyAntsrTransformToImage(tx, img, img)
# plot(img,img2)
}
