% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyAntsrTransformToPixels.R
\name{applyAntsrTransformToPixels}
\alias{applyAntsrTransformToPixels}
\title{applyAntsrTransformToPixels}
\usage{
applyAntsrTransformToPixels(transform, image, pixel.as = "vector",
  in.place = FALSE)
}
\arguments{
\item{transform}{antsrTransform}

\item{image}{antsImage of pixels to transform}

\item{pixel.as}{type of data stored as pixels ("vector", "diffusiontensor")}

\item{in.place}{modify the input image instead of returning a new image}
}
\value{
antsImage
}
\description{
Apply transform to pixel data (vectors, tensors, etc)
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsrTransform", precision="float", type="AffineTransform", dimension=2 )
setAntsrTransformParameters(tx, c(0.9,0,0,1.1,10,11) )
img2 = applyAntsrTransformToImage(tx, img, img)
# plot(img,img2)
}
