% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountImage.R
\name{cellCountImage}
\alias{cellCountImage}
\title{cellCountImage}
\usage{
cellCountImage(mesh, reference)
}
\arguments{
\item{mesh}{'antsrMesh' of polylines}

\item{reference}{antsImage that defines the physical space to consider (points outside of image space are ignored)}
}
\description{
return image that counts the number of unique cells that contain a point in each voxel
}
