% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellEdgeLength.R
\name{cellEdgeLength}
\alias{cellEdgeLength}
\title{cellEdgeLength}
\usage{
cellEdgeLength(mesh, index = NULL)
}
\arguments{
\item{mesh}{'antsrMesh defining a point set and cells}

\item{index}{the cell to examine}
}
\description{
sum of Lengths of cell edges
}
