% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellImageValueSummary.R
\name{cellImageValueSummary}
\alias{cellImageValueSummary}
\title{cellImageValueSummary}
\usage{
cellImageValueSummary(mesh, image, measure = "mean", index = NULL)
}
\arguments{
\item{mesh}{'antsrMesh defining a point set and cells}

\item{image}{an 'antsImage' to average over}

\item{measure}{summary measure to obtain (default="mean")}

\item{index}{the cell/s to examine}
}
\description{
summary of image value sover all points in a cell (mean, max, etc)
}
