% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPointInMask.R
\name{cellPointInMask}
\alias{cellPointInMask}
\title{cellPointInMask}
\usage{
cellPointInMask(mesh, mask, index = NULL)
}
\arguments{
\item{mesh}{'antsrMesh defining a point set and cells}

\item{mask}{an 'antsImage'}

\item{index}{the cell/s to examine}
}
\description{
indicate if a cell has a point inside an image mask
}
