% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionToChannel.R
\name{channelToDimension}
\alias{channelToDimension}
\title{channelToDimension}
\usage{
channelToDimension(img, dimension = (img@dimension + 1), spacing = 1,
  origin = 0, dircol = NULL, dirrow = NULL)
}
\arguments{
\item{img}{a multi component antsImage}

\item{dimension}{the channels are converted to}

\item{spacing}{spacing for new dimension}

\item{origin}{origin for new dimension}

\item{dircol}{column values for direction matrix in new dimension}

\item{dirrow}{row values for direction matrix in new dimension}
}
\description{
convert channels to a dimension
}
