% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicTracking.R
\name{deterministicTracking}
\alias{deterministicTracking}
\title{labelsToPoints}
\usage{
deterministicTracking(img, seeds)
}
\arguments{
\item{img}{a multichannel antsImage of a displacment field}

\item{seeds}{a matrix of seed points for tracking}
}
\description{
return point coordinates of all voxels with specified label
}
