% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionToChannel.R
\name{dimensionToChannel}
\alias{dimensionToChannel}
\title{dimensionToChannel}
\usage{
dimensionToChannel(img, dimension = length(dim(img)))
}
\arguments{
\item{img}{an antsImage}

\item{dimension}{the dimension that is converted to a channel}
}
\description{
create multichannel image by reducing dimension by one
}
