% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiAnisotropy.R
\name{dtiAnisotropy}
\alias{dtiAnisotropy}
\title{dtiAnisotropy}
\usage{
dtiAnisotropy(img, method = "FractionalAnisotropy")
}
\arguments{
\item{img}{a 6-channel antsImage}

\item{method}{the measure to return
\itemize{
  \item{FractionalAnisotropy}{}
  \item{RelativeAnisotropy}{}
}}

\item{method}{which type of anisotropy index to calculate}
}
\description{
calculate an anistropy image from DTI
}
