% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiColorMap.R
\name{dtiColorMap}
\alias{dtiColorMap}
\title{dtiColorMap}
\usage{
dtiColorMap(img, weight = "FractionalAnisotropy")
}
\arguments{
\item{img}{a 6-channel antsImage}

\item{weight}{method used to weight color values'}
}
\description{
calculate an RGB map of primary direction weighted by anisotropy
}
