% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiConvert.R
\name{dtiConvert}
\alias{dtiConvert}
\title{dtiConvert}
\usage{
dtiConvert(img, input, output = "ANTsR")
}
\arguments{
\item{img}{a DT image to convert}

\item{input}{the format of the input image
\itemize{
  \item{ANTsR}{}
  \item{MRtrix}{}
}}

\item{output}{the format of the returned image
\itemize{
  \item{ANTsR}{}
  \item{MRTrix}{}
}}
}
\description{
convert between package formats
}
