% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiEigenSystem.R
\name{dtiEigenSystem}
\alias{dtiEigenSystem}
\title{dtiEigenSystem}
\usage{
dtiEigenSystem(img, method = "itk-svd")
}
\arguments{
\item{img}{a 6-channel antsImage}

\item{method}{method used to calculate the Eigen system, default is 'itk-svd'
\itemize{
  \item{itk-svd}{Uses ITK based call to netlib/vnl SVD. Is faster.}
  \item{eigen}{Uses R's 'eigen' function}
}}
}
\description{
calculate eigen values and vectors
}
