% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiFromVectorImage.R
\name{dtiFromVectorImage}
\alias{dtiFromVectorImage}
\title{dtiFromVectorImage}
\usage{
dtiFromVectorImage(vImg)
}
\arguments{
\item{vImg}{a 6-channel antsImage}
}
\description{
create a diffusion image from a 6-channel antsImage
}
