% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtiReconstruction.R
\name{dtiReconstruction}
\alias{dtiReconstruction}
\title{dtiReconstruction}
\usage{
dtiReconstruction(dwi, gradients, method)
}
\arguments{
\item{method}{the summary meaure to return
\itemize{
  \item{Trace}{Sum of eigenvalues}
  \item{MeanDiffusion}{Average eigenvalue}
  \item{AxialDiffusion}{Largest eigenvalue}
  \item{RadialDiffusion}{Mean of two smallest eigenvalues}
}}

\item{img}{a 6-channel antsImage or and eigen sytem list}

\item{method}{the reconstruction algorithm to use (default = "itk-svd")
\itemize{
  \item{itk-svd}{uses the itk::DiffusionTensor3DReconstructionImageFilter filter}
}}
}
\description{
calculate an anistropy image from DTI
}
