% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInImage.R
\name{indexIsInImage}
\alias{indexIsInImage}
\title{indexIsInImage}
\usage{
indexIsInImage(image, index)
}
\arguments{
\item{image}{'antsImage' defining a physical space}

\item{index}{location to test}
}
\description{
is this index in the image space
}
