% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInImage.R
\name{isInImage}
\alias{isInImage}
\title{isInImage}
\usage{
isInImage(image, coordinate, type = "point")
}
\arguments{
\item{image}{'antsImage' defining a physical space}

\item{coordinate}{point or index to test}

\item{type}{either "point" or "index"}
}
\description{
is this point or index in the image space
}
