% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInMesh.R
\name{isInMesh}
\alias{isInMesh}
\title{isInMesh}
\usage{
isInMesh(mesh, coordinate, tolerance = 1e-05)
}
\arguments{
\item{mesh}{'antsrMesh defining a point set}

\item{coordinate}{point or index to test}

\item{tolerance}{any point with a euclidan distance less than this is considered equal}
}
\description{
is this point or in the mesh
}
