% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiToItk.R
\name{itkToNifti}
\alias{itkToNifti}
\title{itkToNifti}
\usage{
itkToNifti(points, niftiImage = NA, antsImage = NA, filename = NA)
}
\arguments{
\item{points}{the locations of interest}

\item{niftiImage}{an niftiImage}

\item{antsImage}{an antsImage}

\item{filename}{name of image file}

\item{points}{the locations of interest}

\item{type}{'point' or 'index'}

\item{interpolation}{options are: 'linear'}
}
\description{
convert world coordinates from ITK to NIFTI
}
