% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelsToPoints.R
\name{labelsToPoints}
\alias{labelsToPoints}
\title{labelsToPoints}
\usage{
labelsToPoints(img, label = 1)
}
\arguments{
\item{img}{an antsImage}

\item{label}{the value of interest}
}
\description{
return point coordinates of all voxels with specified label
}
