% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiToItk.R
\name{niftiToItk}
\alias{niftiToItk}
\title{niftiToItk}
\usage{
niftiToItk(points, niftiImage = NA, antsImage = NA, filename = NA,
  in.place = FALSE)
}
\arguments{
\item{points}{the locations of interest (matrix or antsrMesh)}

\item{niftiImage}{an niftiImage}

\item{antsImage}{an antsImage}

\item{filename}{name of image file}

\item{in.place}{transform points "in-place", for antsrMesh only}
}
\description{
convert world coordinates from NIFTI to ITK
}
