% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physicalTensorsToIndexTensors.R
\name{physicalTensorsToIndexTensors}
\alias{physicalTensorsToIndexTensors}
\title{physicalTensorsToIndexTensors}
\usage{
physicalTensorsToIndexTensors(image, in.place = FALSE)
}
\arguments{
\item{image}{antsImage of pixels to transform}

\item{in.place}{modify the input image instead of returning a new image}
}
\value{
antsImage
}
\description{
convert image of physical Tensors to index Tensors
}
