% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physicalVectorsToIndexVectors.R
\name{physicalVectorsToIndexVectors}
\alias{physicalVectorsToIndexVectors}
\title{physicalVectorsToIndexVectors}
\usage{
physicalVectorsToIndexVectors(image, in.place = FALSE)
}
\arguments{
\item{image}{antsImage of pixels to transform}

\item{in.place}{modify the input image instead of returning a new image}
}
\value{
antsImage
}
\description{
convert image of physical vectors to index vectors
}
