% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCountImage.R
\name{pointCountImage}
\alias{pointCountImage}
\title{pointCountImage}
\usage{
pointCountImage(points, img)
}
\arguments{
\item{points}{the locations of interest}

\item{img}{an antsImage}

\item{type}{'point' or 'index'}

\item{interpolation}{options are: 'linear'}
}
\description{
return image values at points or indices
}
