% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInImage.R
\name{pointIsInImage}
\alias{pointIsInImage}
\title{pointIsInImage}
\usage{
pointIsInImage(image, point)
}
\arguments{
\item{image}{'antsImage' defining a physical space}

\item{point}{spatial point}
}
\description{
is this point in the image space
}
