% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylineConnections.R
\name{polylineConnections}
\alias{polylineConnections}
\title{polylineConnections}
\usage{
polylineConnections(mesh, regions, seeds = NULL, method = "forward",
  max.regions = 2)
}
\arguments{
\item{img}{an antsImage}

\item{label}{the value of interest}
}
\description{
return point coordinates of all voxels with specified label
}
