% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylineVoxelCount.R
\name{polylineVoxelCount}
\alias{polylineVoxelCount}
\title{polylineVoxelCount}
\usage{
polylineVoxelCount(mesh, reference)
}
\arguments{
\item{reference}{antsImage that defines the physical space to consider}

\item{img}{an antsImage}
}
\description{
return point coordinates of all voxels with specified label
}
