% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{read.antsrMesh}
\alias{read.antsrMesh}
\title{read.antsrMesh}
\usage{
read.antsrMesh(filename, dimension = 3, pixeltype = "float")
}
\arguments{
\item{filename}{name of the file to read}

\item{dimension}{dimension of point data in mesh}

\item{pixeltype}{float or double}
}
\value{
antsrMesh
}
\description{
read a file into an antsrMesh
}
