% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamlineTargetsFromSeeds.R
\name{streamlineTargetsFromSeed}
\alias{streamlineTargetsFromSeed}
\title{streamlineTargetsFromSeed}
\usage{
streamlineTargetsFromSeed(mesh, seeds, image, index)
}
\arguments{
\item{mesh}{'antsrMesh defining a point set and cells (streamlines)}

\item{seeds}{a vector indicating the seed index for each streamline}

\item{image}{an 'antsImage' of labeled regions}

\item{index}{the cell/s to examine}
}
\description{
from seed, find labeled targets
}
