% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{write.antsrMesh}
\alias{write.antsrMesh}
\title{write.antsrMesh}
\usage{
write.antsrMesh(mesh, filename, seeds = NULL, cells.as = "NA")
}
\arguments{
\item{mesh}{antsrMesh to write}

\item{filename}{name of the file to read}

\item{seeds}{seed indices (for Camino files)}

\item{what}{VTK-type of cell should cell data be written as ("NA", "polygon", "line")}
}
\description{
write an antsrMesh to a file
}
