% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\docType{class}
\name{antsrMesh-class}
\alias{antsrMesh-class}
\alias{show,antsrMesh-method}
\alias{initialize,antsrMesh-method}
\alias{initialize,antsrTransform-method}
\title{antsrMesh}
\usage{
\S4method{show}{antsrMesh}(object)

\S4method{initialize}{antsrMesh}(.Object, dimension = 3,
  precision = "float", reserve = 0)
}
\arguments{
\item{object}{input object to convert}

\item{.Object}{input object to convert}

\item{dimension}{dimensionality of the transform (2,3,or 4)}

\item{precision}{string e.g. "float" or "double"}
}
\description{
class for point sets and meshes
}
\section{Slots}{

\describe{
\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{precision}}{math precision is float or double'}

\item{\code{pointer}}{to the memory location of the itk object}
}}

