% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshCreate}
\alias{antsrMeshCreate}
\title{antsrMeshCreate}
\usage{
antsrMeshCreate(dimension = 3, precision = "float", reserve = 0)
}
\arguments{
\item{dimension}{number of dimensions}

\item{precision}{use 'float' or 'double' for values}

\item{reserve}{number of points to allocate on creation}
}
\value{
antsrMesh
}
\description{
create a mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
}
