% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{antsrMeshSetPoint}
\alias{antsrMeshSetPoint}
\title{antsrMeshSetPoint}
\usage{
antsrMeshSetPoint(mesh, point, identifier)
}
\arguments{
\item{mesh}{an 'antsrMesh'}

\item{point}{spatial point to set in mesh}

\item{identifier}{identified for point to set}
}
\description{
set a given point in mesh
}
\examples{
x =  antsrMeshCreate( 3, "float", reserve=128 )
antsrMeshSetPoint( x, c(0,0,0), 1 )
}
