% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrRegistration_class.R
\docType{class}
\name{antsrRegistration-class}
\alias{antsrRegistration-class}
\alias{show,antsrRegistration-method}
\title{antsrRegsitration}
\usage{
\S4method{show}{antsrRegistration}(object)
}
\arguments{
\item{object}{input object to convert}

\item{.Object}{input object to convert}

\item{precision}{string e.g. "float" or "double"}

\item{dimension}{dimensionality of the transform (2,3,or 4)}
}
\description{
C++ type used to represent an ITK image registration filter
}
\section{Slots}{

\describe{
\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{precision}}{math precision is float or double'}

\item{\code{pointer}}{to the memory location of the itk object}
}}

