% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{applyAntsrTransformToMesh}
\alias{applyAntsrTransformToMesh}
\title{applyAntsrTransformToMesh}
\usage{
applyAntsrTransformToMesh(transform, mesh, interpolation = "linear")
}
\arguments{
\item{transform}{antsrTransform}

\item{mesh}{antsrMesh to transform}

\item{interpolation}{type of interpolator to use}
}
\value{
antsrMesh
}
\description{
Apply transform/s to an antsrMesh
}
