% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMesh_class.R
\name{applyAntsrTransformToMesh}
\alias{applyAntsrTransformToMesh}
\title{applyAntsrTransformToMesh}
\usage{
applyAntsrTransformToMesh(transform, image, reference,
  interpolation = "linear")
}
\arguments{
\item{transform}{antsrTransform}

\item{image}{antsrMesh to transform}

\item{interpolation}{type of interpolator to use}
}
\value{
antsImage
}
\description{
Apply transform/s to an antsrMesh
}
