% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicTracking.R
\name{deterministicTracking}
\alias{deterministicTracking}
\title{labelsToPoints}
\usage{
deterministicTracking(img, seeds, mask = NULL)
}
\arguments{
\item{img}{a multichannel antsImage of a displacment field}

\item{seeds}{a matrix of seed points for tracking}

\item{mask}{a mask of brain vs background'}
}
\description{
return point coordinates of all voxels with specified label
}
