% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylineConnections.R
\name{polylineConnections}
\alias{polylineConnections}
\title{polylineConnections}
\usage{
polylineConnections(mesh, regions, seeds = NULL, method = "forward",
  max.regions = 2)
}
\arguments{
\item{mesh}{'antsrMesh' of polylines}

\item{regions}{'antsImage' of labeled regions}

\item{seeds}{vector of seed indices}

\item{method}{method for determining connections}

\item{max.regions}{maximum number of regions that a single polyline can connect}
}
\description{
return point coordinates of all voxels with specified label
}
